scoot175 - renamed daemon_cli to scoot, and refactored the command processing segments into functions.
